- Added User Signups Template with interactive data.

+++ DemoApplication/DemoApplication/Models/Common/BarGraphModel.cs
@@ -11,7 +11,7 @@ namespace DemoApplication.Models.Common
 
         public string Title { get; set; }
 
-        public int MaxValue { get; set; }
+        public decimal MaxValue { get; set; }
 
         public List<BarGraphSection> Sections { get; set; } 
     }
